+++ src/constants/home.constants.ts
@@ -2,7 +2,7 @@ import { BottomNavbarItem } from "@/types/home.types"
 
 const BottomNavbarItemList: BottomNavbarItem[] = [
   { icon: "home_rounded", text: "랜덤매칭", link: "/home" },
-  { icon: "graphic_eq_rounded", text: "매칭방", link: "/home" },
+  { icon: "graphic_eq_rounded", text: "매칭방", link: "/match" },
   { icon: "person_rounded", text: "프로필", link: "/profile" },
   { icon: "settings_rounded", text: "설정", link: "/setting" },
 ]
