+++ src/styles/theme.ts
@@ -1,0 +1,16 @@
+"use client"
+import { createTheme } from "@mui/material"
+
+/* Custom Theme 정의 */
+export const theme = createTheme({
+  palette: {
+    primary: {
+      main: "#5E1EE7",
+    },
+  },
+  typography: {
+    h1: {
+      color: "red",
+    },
+  },
+})
