+++ src/types/common.types.ts
@@ -1,0 +1,7 @@
+export type BottomNavbarKey = "/home" | "match" | "profile" | "/setting"
+
+export type BottomNavbarItem = {
+  icon: string
+  text: string
+  link: string
+}
