+++ src/types/home.types.ts
@@ -1,9 +1,3 @@
-export type BottomNavbarItem = {
-  icon: string
-  text: string
-  link: string
-}
-
 export type TextQueryResult = {
   formattedAddress: string
   displayName: {
